Refactor: renamed sidenavs
